+++ watermark.sh
@@ -1,11 +1,41 @@
 #! /bin/bash
 
+
+# constants
 WATERMARK=/home/kakurady/works/2015/watermark_nekotoba2.png
+
+USAGE_CMDLINE="Usage: $0 [-gk] files ..."
+
+USAGE="$USAGE_CMDLINE
+    -g      set gravity
+    -k      keep original file
+"
+
+#variables
+keep_original=""
+gravity="southeast"
+
+while getopts g:k f
+do
+    case $f in
+    g)      gravity=$OPTARG;;
+    k)      keep_original="y";;
+    \?)     echo "$USAGE"; exit 1;;
+    esac
+done
+shift $(($OPTIND - 1))
 
 if [ ! -f "$WATERMARK" ]
 then 
-	echo "watermark image not found, exiting"
+	echo "watermark image ($WATERMARK) not found, exiting."
+	
 	exit 1
+fi
+
+if [ $# -lt 1 ]
+then
+    echo "$USAGE"
+    exit 1
 fi
 
 if [ ! -d watermarked ] 
@@ -21,11 +51,11 @@ doIt() {
 	echo "working on $1"
 	#convert full-sized image
 	convert $1 "${1%.*}.tga"
-	~/Downloads/mozjpeg/build/cjpeg -quality 90 -targa -outfile "${1%.*}.jpg" "${1%.*}.tga"
+	~/Downloads/mozjpeg/build/cjpeg -quality 92 -targa -outfile "${1%.*}.jpg" "${1%.*}.tga"
 	rm "${1%.*}.tga" 
 	
 	#composite watermarked image
-	composite -gravity southeast -geometry +32+32 "$WATERMARK" "$1" "${1%.*}_watermarked.tga"
+	composite -gravity $gravity -geometry +32+32 "$WATERMARK" "$1" "${1%.*}_watermarked.tga"
 	~/Downloads/mozjpeg/build/cjpeg -quality 70 -targa -outfile "watermarked/${1%.*}.jpg" "${1%.*}_watermarked.tga" 
 	
 	#shrink down watermarked image
@@ -36,11 +66,22 @@ doIt() {
 	
 	#add exif tags
 	exiftool -tagsFromFile "$1" -overwrite_original "${1%.*}.jpg" "resized/${1%.*}.jpg" "watermarked/${1%.*}.jpg"
-	mv "$1.out.pp3" "${1%.*}.jpg.out.pp3"
-	rm $1
+	if [ -f "$1.out.pp3" ]
+	then
+    	mv "$1.out.pp3" "${1%.*}.jpg.out.pp3"
+    fi
+    if [ ! "$keep_original" ]
+    then
+    	rm $1
+    else 
+        true
+        #echo "keep original"
+	fi
 }
 
 #for f in *.tif; do doIt $f; done
 export -f doIt
 export WATERMARK
-parallel --bar doIt ::: *.tif
+export keep_original
+export gravity
+parallel --bar doIt ::: "$@"
